Hide step implementation details from public api
