+++ core/src/main/java/cucumber/api/HookStep.java
@@ -1,13 +1,5 @@
 package cucumber.api;
 
-import cucumber.runtime.DefinitionMatch;
-
-public abstract class HookStep extends Step {
-
-    protected HookStep(DefinitionMatch definitionMatch) {
-        super(definitionMatch);
-    }
-
-    public abstract HookType getHookType();
-
+public interface HookStep extends Step {
+    HookType getHookType();
 }
