+++ core/src/main/java/cucumber/runner/TestCase.java
@@ -1,0 +1,103 @@
+package cucumber.runner;
+
+import cucumber.api.Result;
+import cucumber.api.event.TestCaseFinished;
+import cucumber.api.event.TestCaseStarted;
+import cucumber.runtime.ScenarioImpl;
+import gherkin.events.PickleEvent;
+import gherkin.pickles.PickleLocation;
+import gherkin.pickles.PickleTag;
+
+import java.util.ArrayList;
+import java.util.Collections;
+import java.util.List;
+
+class TestCase implements cucumber.api.TestCase {
+    private final PickleEvent pickleEvent;
+    private final List<PickleTestStep> testSteps;
+    private final boolean dryRun;
+    private final List<Step> beforeHooks;
+    private final List<Step> afterHooks;
+
+    public TestCase(List<PickleTestStep> testSteps, List<Step> beforeHooks, List<Step> afterHooks, PickleEvent pickleEvent, boolean dryRun) {
+        this.testSteps = testSteps;
+        this.beforeHooks = beforeHooks;
+        this.afterHooks = afterHooks;
+        this.pickleEvent = pickleEvent;
+        this.dryRun = dryRun;
+    }
+
+    @Deprecated
+    public TestCase(List<PickleTestStep> testSteps, PickleEvent pickleEvent, boolean dryRun) {
+        this(testSteps, Collections.<Step>emptyList(), Collections.<Step>emptyList(), pickleEvent, dryRun);
+    }
+
+    void run(EventBus bus) {
+        boolean skipNextStep = this.dryRun;
+        Long startTime = bus.getTime();
+        bus.send(new TestCaseStarted(startTime, this));
+        ScenarioImpl scenarioResult = new ScenarioImpl(bus, pickleEvent);
+
+        for (Step before : beforeHooks) {
+            Result stepResult = before.run(bus, pickleEvent.pickle.getLanguage(), scenarioResult, dryRun);
+            skipNextStep |= !stepResult.is(Result.Type.PASSED);
+            scenarioResult.add(stepResult);
+        }
+
+        for (Step step : testSteps) {
+            Result stepResult = step.run(bus, pickleEvent.pickle.getLanguage(), scenarioResult, skipNextStep);
+            skipNextStep |= !stepResult.is(Result.Type.PASSED);
+            scenarioResult.add(stepResult);
+        }
+
+        for (Step after : afterHooks) {
+            Result stepResult = after.run(bus, pickleEvent.pickle.getLanguage(), scenarioResult, dryRun);
+            scenarioResult.add(stepResult);
+        }
+
+        Long stopTime = bus.getTime();
+        bus.send(new TestCaseFinished(stopTime, this, new Result(scenarioResult.getStatus(), stopTime - startTime, scenarioResult.getError())));
+    }
+
+    @Override
+    public List<cucumber.api.Step> getTestSteps() {
+        List<cucumber.api.Step> steps = new ArrayList<cucumber.api.Step>();
+        steps.addAll(beforeHooks);
+        for (PickleTestStep step : testSteps) {
+            steps.addAll(step.getBeforeStepHookSteps());
+            steps.add(step);
+            steps.addAll(step.getAfterStepHookSteps());
+        }
+        steps.addAll(afterHooks);
+        return steps;
+    }
+
+    @Override
+    public String getName() {
+        return pickleEvent.pickle.getName();
+    }
+
+    @Override
+    public String getScenarioDesignation() {
+        return fileColonLine(pickleEvent.pickle.getLocations().get(0)) + " # " + getName();
+    }
+
+    @Override
+    public String getUri() {
+        return pickleEvent.uri;
+    }
+
+    @Override
+    public int getLine() {
+        return pickleEvent.pickle.getLocations().get(0).getLine();
+    }
+
+    private String fileColonLine(PickleLocation location) {
+        return pickleEvent.uri + ":" + location.getLine();
+    }
+
+    @Override
+    public List<PickleTag> getTags() {
+        return pickleEvent.pickle.getTags();
+    }
+}
