+++ core/src/test/java/cucumber/runner/EventBusTest.java
@@ -1,7 +1,6 @@
 package cucumber.runner;
 
 import cucumber.api.Result;
-import cucumber.api.TestStep;
 import cucumber.api.event.EventHandler;
 import cucumber.api.event.TestStepFinished;
 import cucumber.api.event.TestStepStarted;
@@ -16,7 +15,7 @@ public class EventBusTest {
     @Test
     public void handlers_receive_the_events_they_registered_for() {
         EventHandler<TestStepFinished> handler = mock(EventHandler.class);
-        TestStep testStep = mock(TestStep.class);
+        cucumber.api.TestStep testStep = mock(cucumber.api.TestStep.class);
         Result result = mock(Result.class);
         TestStepFinished event = new TestStepFinished(0l, testStep, result);
 
@@ -30,7 +29,7 @@ public class EventBusTest {
     @Test
     public void handlers_do_not_receive_the_events_they_did_not_registered_for() {
         EventHandler handler = mock(EventHandler.class);
-        TestStep testStep = mock(TestStep.class);
+        cucumber.api.TestStep testStep = mock(cucumber.api.TestStep.class);
         TestStepStarted event = new TestStepStarted(0l, testStep);
 
         EventBus bus = new EventBus(new TimeServiceStub(0));
